feature: Add font support for Manifold and klima
These two fonts are the brand fonts that DSA uses.
It makes sense to add in font support for the brand fonts, so I've gone ahead and done that

I've also gone and modified the SCSS files to respect the primary font of Manifold.
This will slightly adjust the layout of the site.